+++ cmd2/decorators.py
@@ -13,7 +13,10 @@ from typing import (
 )
 
 from . import constants
-from .argparse_custom import Cmd2AttributeWrapper
+from .argparse_custom import (
+    Cmd2ArgumentParser,
+    Cmd2AttributeWrapper,
+)
 from .command_definition import (
     CommandFunc,
     CommandSet,
@@ -184,19 +187,19 @@ def with_argument_list(
     return arg_decorator
 
 
-#: Function signatures for command functions that use an argparse.ArgumentParser to process user input
+#: Function signatures for command functions that use a Cmd2ArgumentParser to process user input
 #: and optionally return a boolean
 ArgparseCommandFuncOptionalBoolReturn: TypeAlias = Callable[[CmdOrSet, argparse.Namespace], bool | None]
 ArgparseCommandFuncWithUnknownArgsOptionalBoolReturn: TypeAlias = Callable[
     [CmdOrSet, argparse.Namespace, list[str]], bool | None
 ]
 
-#: Function signatures for command functions that use an argparse.ArgumentParser to process user input
+#: Function signatures for command functions that use a Cmd2ArgumentParser to process user input
 #: and return a boolean
 ArgparseCommandFuncBoolReturn: TypeAlias = Callable[[CmdOrSet, argparse.Namespace], bool]
 ArgparseCommandFuncWithUnknownArgsBoolReturn: TypeAlias = Callable[[CmdOrSet, argparse.Namespace, list[str]], bool]
 
-#: Function signatures for command functions that use an argparse.ArgumentParser to process user input
+#: Function signatures for command functions that use a Cmd2ArgumentParser to process user input
 #: and return nothing
 ArgparseCommandFuncNoneReturn: TypeAlias = Callable[[CmdOrSet, argparse.Namespace], None]
 ArgparseCommandFuncWithUnknownArgsNoneReturn: TypeAlias = Callable[[CmdOrSet, argparse.Namespace, list[str]], None]
@@ -213,17 +216,17 @@ ArgparseCommandFunc: TypeAlias = (
 
 
 def with_argparser(
-    parser: argparse.ArgumentParser  # existing parser
-    | Callable[[], argparse.ArgumentParser]  # function or staticmethod
-    | Callable[[CmdOrSetClass], argparse.ArgumentParser],  # Cmd or CommandSet classmethod
+    parser: Cmd2ArgumentParser  # existing parser
+    | Callable[[], Cmd2ArgumentParser]  # function or staticmethod
+    | Callable[[CmdOrSetClass], Cmd2ArgumentParser],  # Cmd or CommandSet classmethod
     *,
     ns_provider: Callable[..., argparse.Namespace] | None = None,
     preserve_quotes: bool = False,
     with_unknown_args: bool = False,
 ) -> Callable[[ArgparseCommandFunc[CmdOrSet]], RawCommandFuncOptionalBoolReturn[CmdOrSet]]:
-    """Decorate a ``do_*`` method to populate its ``args`` argument with the given instance of argparse.ArgumentParser.
+    """Decorate a ``do_*`` method to populate its ``args`` argument with the given instance of Cmd2ArgumentParser.
 
-    :param parser: instance of ArgumentParser or a callable that returns an ArgumentParser for this command
+    :param parser: instance of Cmd2ArgumentParser or a callable that returns an Cmd2ArgumentParser for this command
     :param ns_provider: An optional function that accepts a cmd2.Cmd or cmd2.CommandSet object as an argument and returns an
                         argparse.Namespace. This is useful if the Namespace needs to be prepopulated with state data that
                         affects parsing.
@@ -347,9 +350,9 @@ def with_argparser(
 def as_subcommand_to(
     command: str,
     subcommand: str,
-    parser: argparse.ArgumentParser  # existing parser
-    | Callable[[], argparse.ArgumentParser]  # function or staticmethod
-    | Callable[[CmdOrSetClass], argparse.ArgumentParser],  # Cmd or CommandSet classmethod
+    parser: Cmd2ArgumentParser  # existing parser
+    | Callable[[], Cmd2ArgumentParser]  # function or staticmethod
+    | Callable[[CmdOrSetClass], Cmd2ArgumentParser],  # Cmd or CommandSet classmethod
     *,
     help: str | None = None,  # noqa: A002
     aliases: Sequence[str] | None = None,
@@ -359,7 +362,7 @@ def as_subcommand_to(
 
     :param command: Command Name. Space-delimited subcommands may optionally be specified
     :param subcommand: Subcommand name
-    :param parser: instance of ArgumentParser or a callable that returns an ArgumentParser for this subcommand
+    :param parser: instance of Cmd2ArgumentParser or a callable that returns an Cmd2ArgumentParser for this subcommand
     :param help: Help message for this subcommand which displays in the list of subcommands of the command we are adding to.
                  This is passed as the help argument to subparsers.add_parser().
     :param aliases: Alternative names for this subcommand. This is passed as the alias argument to